refactor(blog): clean up code and improve consistency
- Remove unused imports and dead code across all blog files
- Remove view counts and like buttons from dashboard blog pages
- Clean up debug console.logs and unused functions
- Standardize background colors (bg-background for dashboard, normal for admin)
- Replace alert() with toast.error() for consistent UX
- Move blog features from clubs/pages/blog to dedicated features/blog directory
- Clean unused imports: Eye, Heart, User, Clock, etc.
- Remove hardcoded view counts (247 views, 156 views)
- Improve code maintainability and consistency
